requirements.txt
xxhash
wheel
twilio
pyyaml
flask
gunicorn
pyrebase
pytest
pytz
plan
textblob
plotly
